LayoutTests/inspector/debugger/scripts-panel-expected.txt

Tests that scripts panel UI elements work as intended.


Running: testInitialLoad
Dumping ScriptsNavigator 'Scripts' tab:
  bar.js
  baz.js
  foo.js
  foobar.js
Dumping ScriptsNavigator 'Content scripts' tab:
Source requested for baz.js

Running: testReset
Source requested for baz.js
Dumping ScriptsNavigator 'Scripts' tab:
  bar.js
  baz.js
  foo.js
Dumping ScriptsNavigator 'Content scripts' tab:
Dumping ScriptsNavigator 'Scripts' tab:
  bar.js
Dumping ScriptsNavigator 'Content scripts' tab:

Running: testSourceReplaced
Dumping ScriptsNavigator 'Scripts' tab:
  compiled.js
  foo.js
Dumping ScriptsNavigator 'Content scripts' tab:
Dumping ScriptsNavigator 'Scripts' tab:
  foo.js
  source.js
Dumping ScriptsNavigator 'Content scripts' tab:
